+++ frontend/src/webview/src/types/theme-interface.ts
@@ -1,0 +1,32 @@
+
+import "@mui/material/styles";
+declare module "@mui/material/styles" {
+  interface Palette {
+    shadowColor: string;
+  }
+  interface PaletteOptions {
+    shadowColor?: string;
+    background?: Partial<TypeBackground>;
+  }
+  interface Theme {
+    gradients: {
+      primary: string;
+      secondary: {
+        main: string;
+        light: string;
+      };
+    };
+  }
+  interface ThemeOptions {
+    gradients?: {
+      primary?: string;
+      secondary?: {
+        main?: string;
+        light?: string;
+      };
+    };
+  }
+  interface TypeBackground {
+    nav: string;
+  }
+}
